webapp/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiosqlite>=0.20.0
SQLAlchemy>=2.0.0
PyJWT>=2.8.0
pydantic>=2.0.0
